+++ theming/ui-shadcn/src/theme-classic/_compiled/scheduler-views.tsx
@@ -68,10 +68,10 @@ export function SchedulerViews({
             'items-center',
           ),
           rowEventBeforeClass: (info) => cn(
-            !info.isStartResizable && `${continuationArrowClass} border-e-[5px] border-e-black`
+            !info.isStart && `${continuationArrowClass} border-e-[5px] border-e-black`
           ),
           rowEventAfterClass: (info) => cn(
-            !info.isEndResizable && `${continuationArrowClass} border-s-[5px] border-s-black`
+            !info.isEnd && `${continuationArrowClass} border-s-[5px] border-s-black`
           ),
           rowEventInnerClass: (info) => (
             info.options.eventOverlap