+++ standard/CHANGELOG.md
@@ -6,6 +6,7 @@
 - FIX: virtualization: Now indicator not shown (#8053)
 - FIX: virtualization: scrollLeft JS action will not render timeline (#8054)
 - FIX: Event color not changing when changed via Event::setProp (#8064)
+- FIX: Classic theme timeline event arrows appearing unnecessarily (#8055)
 
 
 ## v7.0.0-rc.1